Fix regitstration

+++ beauty_salon/views.py
@@ -34,20 +34,27 @@ def index(request):
         'masters': masters,
         'comments': comments,
     }
-    if request.method == 'POST':
-        print(request.POST)
-        phone_number = request.POST.get('tel2')
+    if request.method == 'POST' and not 'num1' in request.POST:
+        body = json.loads(request.body)
+        phone_number = request.session['phone_number'] = body['phone_number']
         user, _ = CustomUser.objects.get_or_create(phone_number=phone_number)
-        code = SMSCode.objects.create(number='1234', client=user)
-        print(code.number)
-        code_text = request.POST.get(
-            'num1') + request.POST.get('num2') + request.POST.get('num3') + request.POST.get('num4')
-        print(code_text)
+        SMSCode.objects.filter(client=user).delete()
+        SMSCode.objects.create(number='1234', client=user)
+    if request.method == 'POST' and 'num1' in request.POST:
+        user = CustomUser.objects.get(phone_number=request.session['phone_number'])
+        code = SMSCode.objects.get(client=user)
+        code_text = (
+            request.POST.get('num1')
+            + request.POST.get('num2')
+            + request.POST.get('num3')
+            + request.POST.get('num4')
+        )
         if code_text == code.number:
             login(request, user, backend=settings.AUTHENTICATION_BACKENDS[0])
             code.delete()
         else:
             print('Wrong code')
+            code.delete()
 
     return render(request, 'index.html', context)
